Create chunk_size module
Move everything around the chunk sizing domain into its own module.
